+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added ability to order by an association column
 - Implemented the `action` option for the `Policy` macro
 - Implemented the `Where` filter
+- Implemented the `Scope` filter
 
 ### Changed